git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@207 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/main.cpp
@@ -141,7 +141,7 @@ int main(int argc, char *argv[])
 			oldtime = tLX->fCurTime;
 			
 			// cap the delta
-			tLX->fDeltaTime = MIN(tLX->fDeltaTime, (float)20);
+			tLX->fDeltaTime = MIN(tLX->fDeltaTime, 20.0f);
 
 			// Main frame
 			GameLoop();
